refactor(cli): extract program command registration
